FIX: Modal methods should be wired up in `setupTests` not in module
